Skip type alias addition when a type specifier has been found.
When a type specifier is found for a type alias the alias itself should not be added as the traversal code for the underlying specifier will find this type alias as the parent type and use it to register the class or enum. This way the underlying type
will be visible under the name of the type alias.