Fix crash in setter function in case an additional checker function is used to verify the type conversion of input argument.
This was causing a string-based setter to crash as the generated conditional expression was incorrect.

The original code contains a '!' character before the conversion check. This was fine for type checks with just one expression. In case the expression consisted of a
conjunction of two checks, the negation was incorrectly applied only to the first of them. This caused the second expression not to be evaluated in case the first
expression returned true. Since the second expression also assigned a value to the 'pythonToCpp' pointer, the pointer variable was left uninitialized.

The fix adds an extra pair of brackets around the type check expression to ensure proper expression evaluation.